WealthWise.AI/requirements.txt
# FinGPT Financial AI Services - FastAPI Application
# Complete dependencies for AI-powered financial advisory APIs

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python_jose

# FinGPT and AI/ML Dependencies
torch>=2.1.0
transformers>=4.35.0
peft>=0.6.0
sentencepiece>=0.1.99
accelerate>=0.24.0
bitsandbytes>=0.41.0
datasets>=2.14.0

# HuggingFace Hub Integration
huggingface-hub>=0.17.0

# Financial Data APIs and Processing
yfinance>=0.2.18
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# HTTP Clients and API Integration
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# Data Validation and Serialization
pydantic>=2.4.0
pydantic-settings>=2.0.0
json-repair
python-dateutil

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-jwt>=4.0.0

# Environment and Configuration
python-dotenv>=1.0.0
python-decouple>=3.8

# Caching and Performance
redis>=5.0.0
diskcache>=5.6.0

# Logging and Monitoring
structlog>=23.2.0
colorama>=0.4.6

# Testing (Development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-fastapi-deps>=0.2.0
httpx>=0.25.0
